Move internal packages under internal/
In order to provide a good management API, we should start with only
exposing the parts of the public API that should be public. This
commit moves everything that isn't the public API under internal/ in
line with other Infratographer services.
