+++ internal/celutils/const.go
