+++ internal/celutils/doc.go
