+++ internal/celutils/errors.go
