+++ internal/celutils/utils.go
