+++ internal/rfc8693/errors.go
